[FB] Dev | refactor to use vite
