+++ browser/base/content/floorp-downloadbar.js
