+++ browser/base/content/floorp-pinned-tabs-title.js
