+++ browser/base/content/floorp-preferences.js
@@ -1,0 +1,218 @@
+/* This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
+
+// I glared at the source code for about 3 hours, but for some reason I decided to use the server because it would be unclear because of the Floorp interface settings. God forgive me
+
+{
+  const BROWSER_CHROME_SYSTEM_COLOR = Services.prefs.getIntPref(
+    "floorp.chrome.theme.mode"
+  );
+
+  switch (BROWSER_CHROME_SYSTEM_COLOR) {
+    case 1:
+      Services.prefs.setIntPref("ui.systemUsesDarkTheme", 1);
+      break;
+    case 0:
+      Services.prefs.setIntPref("ui.systemUsesDarkTheme", 0);
+      break;
+    case -1:
+      Services.prefs.clearUserPref("ui.systemUsesDarkTheme");
+      break;
+  }
+}
+
+Services.prefs.addObserver("floorp.chrome.theme.mode", () => {
+  const BROWSER_CHROME_SYSTEM_COLOR = Services.prefs.getIntPref(
+    "floorp.chrome.theme.mode"
+  );
+  switch (BROWSER_CHROME_SYSTEM_COLOR) {
+    case 1:
+      Services.prefs.setIntPref("ui.systemUsesDarkTheme", 1);
+      break;
+    case 0:
+      Services.prefs.setIntPref("ui.systemUsesDarkTheme", 0);
+      break;
+    case -1:
+      Services.prefs.clearUserPref("ui.systemUsesDarkTheme");
+      break;
+  }
+});
+
+// UserAgent
+/*
+0: Default
+1: Chrome (Windows)
+2: Chrome (macOS)
+3: Chrome (Linux)
+4: Chrome (iOS)
+*/
+Services.scriptloader.loadSubScript(
+  "chrome://browser/content/ua_data.js",
+  this
+);
+const BROWSER_SETED_USERAGENT_PREF = "floorp.browser.UserAgent";
+const GENERAL_USERAGENT_OVERRIDE_PREF = "general.useragent.override";
+{
+  const setUserAgent = function (BROWSER_SETED_USERAGENT) {
+    switch (BROWSER_SETED_USERAGENT) {
+      case 0:
+        Services.prefs.clearUserPref(GENERAL_USERAGENT_OVERRIDE_PREF);
+        break;
+      case 1:
+        Services.prefs.setStringPref(
+          GENERAL_USERAGENT_OVERRIDE_PREF,
+          CHROME_STABLE_UA.win
+        );
+        break;
+      case 2:
+        Services.prefs.setStringPref(
+          GENERAL_USERAGENT_OVERRIDE_PREF,
+          CHROME_STABLE_UA.mac
+        );
+        break;
+      case 3:
+        Services.prefs.setStringPref(
+          GENERAL_USERAGENT_OVERRIDE_PREF,
+          CHROME_STABLE_UA.linux
+        );
+        break;
+      case 4:
+        Services.prefs.setStringPref(
+          GENERAL_USERAGENT_OVERRIDE_PREF,
+          "Mozilla/5.0 (iPhone; CPU iPhone OS 16_3 like Mac OS X) AppleWebKit/605.1.15 (KHTML, like Gecko) CriOS/110.0.5481.83 Mobile/15E148 Safari/604.1"
+        );
+        break;
+      case 5:
+        Services.prefs.setStringPref(
+          GENERAL_USERAGENT_OVERRIDE_PREF,
+          Services.prefs.getCharPref("floorp.general.useragent.override")
+        );
+    }
+  };
+
+  {
+    const BROWSER_SETED_USERAGENT = Services.prefs.getIntPref(
+      BROWSER_SETED_USERAGENT_PREF
+    );
+    setUserAgent(BROWSER_SETED_USERAGENT);
+  }
+
+  Services.prefs.addObserver(BROWSER_SETED_USERAGENT_PREF, function () {
+    const BROWSER_SETED_USERAGENT = Services.prefs.getIntPref(
+      BROWSER_SETED_USERAGENT_PREF
+    );
+    setUserAgent(BROWSER_SETED_USERAGENT);
+  });
+}
+
+// backup Floorp Notes Pref
+
+const FLOORP_NOTES_PREF = "floorp.browser.note.memos";
+const FLOORP_NOTES_LATEST_BACKUP_TIME_PREF =
+  "floorp.browser.note.backup.latest.time";
+
+const backupFloorpNotes = async () => {
+  const memos = Services.prefs.getCharPref(FLOORP_NOTES_PREF).slice(1, -1);
+  const time = new Date().getTime();
+  const backup = { [time]: memos };
+  const jsonToStr = JSON.stringify(backup).slice(1, -1);
+  Services.prefs.setCharPref(FLOORP_NOTES_LATEST_BACKUP_TIME_PREF, time);
+
+  try {
+    IOUtils.exists(
+      PathUtils.join(
+        Services.dirsvc.get("ProfD", Ci.nsIFile).path,
+        "floorp_notes_backup.json"
+      )
+    ).then(data => {
+      if (!data) {
+        const backupFilePath = PathUtils.join(
+          Services.dirsvc.get("ProfD", Ci.nsIFile).path,
+          "floorp_notes_backup.json"
+        );
+        IOUtils.writeUTF8(backupFilePath, `{"data":{${jsonToStr},`);
+      } else {
+        const backupFilePath = PathUtils.join(
+          Services.dirsvc.get("ProfD", Ci.nsIFile).path,
+          "floorp_notes_backup.json"
+        );
+        IOUtils.readUTF8(backupFilePath).then(content => {
+          const appText = `${content}${jsonToStr},`;
+          IOUtils.writeUTF8(backupFilePath, appText);
+        });
+      }
+    });
+  } catch (e) {}
+};
+
+if (!Services.prefs.prefHasUserValue(FLOORP_NOTES_PREF)) {
+  backupFloorpNotes();
+}
+
+function getAllBackupedNotes() {
+  const filePath = PathUtils.join(
+    Services.dirsvc.get("ProfD", Ci.nsIFile).path,
+    "floorp_notes_backup.json"
+  );
+  const content = IOUtils.readUTF8(filePath).then(_content => {
+    _content = _content.slice(0, -1) + "}}";
+    return JSON.parse(_content);
+  });
+  return content;
+}
+
+//Backup Limit is 10.
+getAllBackupedNotes().then(content => {
+  const backupLimit = 10;
+  const dataKeys = Object.keys(content.data);
+
+  if (dataKeys.length > backupLimit) {
+    const sortedKeys = dataKeys.sort((a, b) => b - a);
+    const deleteKeys = sortedKeys.slice(backupLimit);
+
+    deleteKeys.forEach(key => {
+      delete content.data[key];
+    });
+
+    const jsonToStr = JSON.stringify(content).slice(0, -2) + ",";
+    const filePath = PathUtils.join(
+      Services.dirsvc.get("ProfD", Ci.nsIFile).path,
+      "floorp_notes_backup.json"
+    );
+    IOUtils.writeUTF8(filePath, jsonToStr);
+  }
+});
+
+/*------------------------------------- user.js -------------------------------------*/
+// Apply Floorp user.js
+
+const FLOORP_USERJS_PREF = "floorp.browser.userjs";
+
+async function applyUserJSCustomize() {
+  const userjsUtils = ChromeUtils.importESModule(
+    "resource:///modules/userjsUtils.sys.mjs"
+  );
+  const pref = Services.prefs.getStringPref("floorp.user.js.customize", "");
+
+  if (pref != "") {
+    const url = userjsUtils.userJsList[pref][0];
+    const PROFILE_DIR = Services.dirsvc.get("ProfD", Ci.nsIFile).path;
+    const userjs = PathUtils.join(PROFILE_DIR, "user.js");
+
+    try {
+      userjs.remove(false);
+    } catch (e) {}
+
+    fetch(url)
+      .then(response => response.text())
+      .then(async data => {
+        const encoder = new TextEncoder("UTF-8");
+        const writeData = encoder.encode(data);
+
+        await IOUtils.write(userjs, writeData);
+      });
+  }
+}
+
+applyUserJSCustomize();
