+++ browser/base/jar.mn
@@ -6,33 +6,30 @@ browser.jar:
 %  content browser %content/browser/ contentaccessible=yes
 
 # Floorp build
-  content/browser/browser-UI-custom.js          (content/browser-UI-custom.js)
-  content/browser/browser-UI-customizing-menu.js (content/browser-UI-customizing-menu.js)
-  content/browser/browser-design.js             (content/browser-design.js)
-  content/browser/browser-downloadbar.js        (content/browser-downloadbar.js)
-  content/browser/browser-favicon-color.js      (content/browser-favicon-color.js)
-  content/browser/browser-tabbar.js             (content/browser-tabbar.js)
-  content/browser/browser-workspaces.js         (content/browser-workspaces.js)
-  content/browser/browser-browserActions.js     (content/browser-browserActions.js)
-  content/browser/browser-pageActions-floorp.js        (content/browser-pageActions.js)
-  content/browser/browser-pinned-tabs-title.js  (content/browser-pinned-tabs-title.js)
-  content/browser/browser-rest-mode.js          (content/browser-rest-mode.js)
-  content/browser/browser-splitView.js          (content/browser-splitView.js)
-  content/browser/browser-chromeCSS.js          (content/browser-chromeCSS.js)
-  content/browser/browser-custom-keyboard-shortcut.js  (content/browser-custom-keyboard-shortcut.js)
+#  content/browser/browser-UI-custom.js          (content/browser-UI-custom.js)
+#  content/browser/browser-UI-customizing-menu.js (content/browser-UI-customizing-menu.js)
+#  content/browser/browser-design.js             (content/browser-design.js)
+#  content/browser/browser-downloadbar.js        (content/browser-downloadbar.js)
+# content/browser/browser-favicon-color.js      (content/browser-favicon-color.js)
+#  content/browser/browser-tabbar.js             (content/browser-tabbar.js)
+#  content/browser/browser-browserActions.js     (content/browser-browserActions.js)
+#  content/browser/floorp-pageActions.js        (content/floorp-pageActions.js)
+#  content/browser/browser-pinned-tabs-title.js  (content/browser-pinned-tabs-title.js)
+#  content/browser/browser-rest-mode.js          (content/browser-rest-mode.js)
+#  content/browser/browser-splitView.js          (content/browser-splitView.js)
+#  content/browser/browser-chromeCSS.js          (content/browser-chromeCSS.js)
+#  content/browser/browser-custom-keyboard-shortcut.js  (content/browser-custom-keyboard-shortcut.js)
   content/browser/browser-verticaltabs.css      (content/resources/css/browser-verticaltabs.css)
-  content/browser/browser-verticaltabs.js       (content/browser-verticaltabs.js)
+#  content/browser/browser-verticaltabs.js       (content/browser-verticaltabs.js)
   content/browser/qr-code-styling/qr-code-styling.js  (library/qr-code-styling/qr-code-styling@1.5.0.js)
-  content/browser/browser-manager-sidebar.js    (content/browser-manager-sidebar.js)
+#  content/browser/browser-manager-sidebar.js    (content/browser-manager-sidebar.js)
   content/browser/browser-floorp.css            (content/resources/css/browser-floorp.css)
-  content/browser/browser-command.js            (content/browser-command.js)
-  content/browser/browser-flex-order.js         (content/browser-flex-order.js)
-  content/browser/browser-overrides.js          (content/browser-overrides.js)
-  content/browser/browser-preferences.js        (content/browser-preferences.js)
-  content/browser/browser-private-container.js  (content/browser-private-container.js)
-  content/browser/browser-statusbar.js          (content/browser-statusbar.js)
-  content/browser/browser-ssb-manager.js        (content/browser-ssb-manager.js)
-  content/browser/browser-context-menu.js       (content/browser-context-menu.js)
+#  content/browser/browser-flex-order.js         (content/browser-flex-order.js)
+#  content/browser/browser-preferences.js        (content/browser-preferences.js)
+#  content/browser/browser-private-container.js  (content/browser-private-container.js)
+#  content/browser/browser-statusbar.js          (content/browser-statusbar.js)
+#  content/browser/browser-ssb-manager.js        (content/browser-ssb-manager.js)
+#  content/browser/browser-context-menu.js       (content/browser-context-menu.js)
   content/browser/browser-bms-window.css        (content/resources/css/browser-bms-window.css)
   content/browser/BMS-extension-needs-white-bg.txt  (sidebar-extension/BMS-extension-needs-white-bg.txt)
   content/browser/ua_data.js                    (content/ua_data.js)
